# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "get": {
        "description": "Returns users. Superuser/fleet manager can view all. Other roles use this for profile loading in frontend.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name, email, or role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users returned successfully"
          }
        },
        "summary": "List users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Only superuser can create users.",
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (e.g. SU-001)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "description": "Superuser/fleet manager can update users. Other roles can update their own profile.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "description": "Approve, reject, or suspend a user.",
        "operationId": "UsersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        },
        "summary": "Update user status",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/vehicles": {
      "get": {
        "operationId": "VehiclesController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status (active, maintenance, blocked, on trip)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vehicles"
          }
        },
        "summary": "List all vehicles",
        "tags": [
          "Vehicles"
        ]
      },
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created"
          }
        },
        "summary": "Add a new vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
    "/api/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get vehicle by ID",
        "tags": [
          "Vehicles"
        ]
      },
      "put": {
        "operationId": "VehiclesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a vehicle",
        "tags": [
          "Vehicles"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Vehicle ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a vehicle",
        "tags": [
          "Vehicles"
        ]
      }
    },
...
  },
  "info": {
    "title": "DeliverSync API",
    "description": "REST API for the DeliverSync delivery management platform. Role-based access is controlled via the `x-user-role` request header. Supported roles: superuser, fleet-manager, business-client, driver.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "john@deliverysync.com"
          },
          "password": {
            "type": "string",
            "example": "Pass@123",
            "description": "Min 8 chars, 1 uppercase, 1 special char"
          },
          "role": {
            "type": "string",
            "example": "fleet-manager",
            "enum": [
              "fleet-manager",
              "business-client",
              "driver",
              "superuser"
            ]
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Pending",
              "Suspended"
            ]
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "companyName": {
            "type": "string",
            "example": "Acme Logistics Inc."
          },
          "company": {
            "type": "string",
            "example": "Acme Logistics Inc."
          },
          "address": {
            "type": "string",
            "example": "123 Business Avenue"
          },
          "profileDetails": {
            "type": "object",
            "description": "Role-specific profile details"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@deliverysync.com"
          },
          "password": {
            "type": "string",
            "example": "Pass@123"
          },
          "role": {
            "type": "string",
            "example": "fleet-manager"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "companyName": {
            "type": "string",
            "example": "Acme Logistics Inc."
          },
          "company": {
            "type": "string",
            "example": "Acme Logistics Inc."
          },
          "address": {
            "type": "string",
            "example": "123 Business Avenue"
          },
          "profileDetails": {
            "type": "object"
          },
          "notifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Suspended",
              "Rejected",
              "Pending Approval"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "TN09AB1234"
          },
          "type": {
            "type": "string",
            "example": "Mini Truck",
            "enum": [
              "Mini Truck",
              "Van",
              "Truck",
              "Bike",
              "SUV",
              "Cargo Van"
            ]
          },
          "capacity": {
            "type": "string",
            "example": "2 Tons"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "On Trip",
              "Maintenance",
              "Blocked"
            ]
          },
          "assignedDriver": {
            "type": "string",
            "example": "Raghav Reddy"
          },
          "maintenance": {
            "type": "string",
            "example": "2026-05-03",
            "description": "Last maintenance date in YYYY-MM-DD format"
          },
          "lastMaintenance": {
            "type": "string",
            "example": "2026-05-03",
            "description": "Alias for maintenance date"
          }
        },
        "required": [
          "plate",
          "type"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string",
            "example": "TN09AB1234"
          },
          "type": {
            "type": "string",
            "example": "Van",
            "enum": [
              "Mini Truck",
              "Van",
              "Truck",
              "Bike",
              "SUV",
              "Cargo Van"
            ]
          },
          "capacity": {
            "type": "string",
            "example": "1 Ton"
          },
          "status": {
            "type": "string",
            "example": "Maintenance",
            "enum": [
              "Active",
              "On Trip",
              "Maintenance",
              "Blocked"
            ]
          },
          "assignedDriver": {
            "type": "string",
            "example": "Kiran Teja"
          },
          "maintenance": {
            "type": "string",
            "example": "2026-05-03",
            "description": "Last maintenance date in YYYY-MM-DD format"
          },
          "lastMaintenance": {
            "type": "string",
            "example": "2026-05-03",
            "description": "Alias for maintenance date"
          }
        }
      },
...
    }
  }
}